front: move handleTrainDrag in a separate file
