+++ front/src/applications/operationalStudies/views/Scenario/components/SimulationResults/SimulationResults.tsx
@@ -10,6 +10,7 @@ import useSimulationResults from 'applications/operationalStudies/hooks/useSimul
 import type { Board } from 'applications/operationalStudies/types';
 import { type Conflict } from 'common/api/osrdEditoastApi';
 import SimulationWarpedMap from 'common/Map/WarpedMap/SimulationWarpedMap';
+import createHandleTrainDrag from 'modules/simulationResult/components/SpaceTimeChartWrapper/helpers/createHandleTrainDrag';
 import SpaceTimeChartWrapper, {
   MANCHETTE_WITH_SPACE_TIME_CHART_DEFAULT_HEIGHT,
 } from 'modules/simulationResult/components/SpaceTimeChartWrapper/SpaceTimeChartWrapper';
@@ -23,11 +24,11 @@ import type { ProjectionData } from 'modules/simulationResult/types';
 import TimesStopsOutput from 'modules/timesStops/TimesStopsOutput';
 import { findExceptionWithOccurrenceId } from 'modules/timetableItem/helpers/pacedTrain';
 import type { TimetableItemWithDetails } from 'modules/timetableItem/types';
-import type { TimetableItemId, TrainId } from 'reducers/osrdconf/types';
+import type { TimetableItemId } from 'reducers/osrdconf/types';
 import { toggleDisplayOnlyPathSteps, updateSelectedTrainId } from 'reducers/simulationResults';
 import {
-  getTrainIdUsedForProjection,
   getDisplayOnlyPathSteps,
+  getTrainIdUsedForProjection,
 } from 'reducers/simulationResults/selectors';
 import { useAppDispatch } from 'store';
 import {
@@ -145,55 +146,12 @@ const SimulationResults = ({
     return exception?.summary ?? pacedTrain.summary;
   }, [timetableItemsWithDetails, selectedTrainId]);
 
-  const handleTrainDrag = async ({
-    draggedTrainId,
-    newDepartureTime,
-    initialDepartureTime,
-    stopPanning,
-  }: {
-    draggedTrainId: TrainId;
-    newDepartureTime: Date;
-    initialDepartureTime: Date;
-    stopPanning: boolean;
-  }) => {
-    const draggedTimetatbleItemId = isTrainScheduleId(draggedTrainId)
-      ? draggedTrainId
-      : extractPacedTrainIdFromOccurrenceId(draggedTrainId);
-    const draggedTrain = projectPathTrainResult.find(
-      (train) => train.id === draggedTimetatbleItemId
-    );
-    if (!draggedTrain) return;
-
-    const newTrainData = { ...draggedTrain, departureTime: newDepartureTime };
-
-    // Handle updating track occupancy data (with no distant update yet, so with stopPanning: false)
-    await handleTrainDragInTrackOccupancy({
-      draggedTrainId,
-      stopPanning: false,
-      initialDepartureTime,
-      newTrainData,
-    });
-
-    if (stopPanning) {
-      // update in the database
-      await updateTrainDepartureTime(draggedTimetatbleItemId, newDepartureTime);
-
-      // Handle retrieving track occupancy data from server (so with stopPanning: true):
-      await handleTrainDragInTrackOccupancy({
-        draggedTrainId,
-        stopPanning,
-        initialDepartureTime,
-        newTrainData,
-      });
-    } else {
-      // update in the state
-      setProjectPathTrainResult(
-        projectPathTrainResult.map((train) =>
-          train.id === draggedTimetatbleItemId ? newTrainData : train
-        )
-      );
-    }
-  };
+  const handleTrainDrag = createHandleTrainDrag({
+    projectPathTrainResult,
+    setProjectPathTrainResult,
+    handleTrainDragInTrackOccupancy,
+    updateTrainDepartureTime,
+  });
 
   const { etcsBrakingCurves, fetchEtcsBrakingCurves } = useEtcsBrakingCurves(
     simulationResults?.rollingStock?.etcs_brake_params !== null,
